+++ src/app/core/default-values/x-inherited-value/inherited-value-button/inherited-value-button.component.html
